fix: remove unnecessary cast in quick_note_store.dart

+++ lib/services/quick_note_store.dart
@@ -307,7 +307,7 @@ class QuickNoteStore {
     final index = _autoSyncFailureCount.clamp(
       0,
       _quickNoteAutoSyncDelays.length - 1,
-    ) as int;
+    );
     return _quickNoteAutoSyncDelays[index];
   }
